pyproject.toml
[tool.poetry]
name = "scrapy-zap"
version = "0.1.0"
description = "raspando dados do zap imoveis"
authors = ["gabrielpereira12345 <gabrielpereira00PereiraGabriel@outlook.com>"]
readme = "README.md"
packages = [{include = "scrapy_zap"}]

[tool.poetry.dependencies]
python = "^3.10"
w3lib = ">=2.1.1,<2.2.0"
unidecode = ">=1.3.6,<1.4.0"
scrapeops-scrapy-proxy-sdk = "^1.0"
attrs = ">=19.1.0"
websockets = "^11.0.3"
asyncio = ">=3.4.3,<3.5.0"
scrapy-playwright = ">=0.0.27,<0.1.0"
pylint = "^2.17.5"
geopy = ">=2.3.0,<2.4.0"
pandas = ">=2.0.3,<2.1.0"
proxybroker2 = ">=2.0.0a4,<2.1.0"
scikit-learn = ">=1.5.0,<1.6.0"
lightgbm = ">=4.3.0,<4.4.0"
seaborn = ">=0.13.2,<0.14.0"
optuna = ">=3.6.1,<3.7.0"
ipywidgets = "^8.1.3"
plotly = ">=5.22.0,<5.23.0"
nbformat = ">=5.10.4,<5.11.0"
numpy = ">=1.26.4,<1.27.0"
xgboost = ">=2.1.1,<2.2.0"
nbclient = ">=0.10.0,<0.11.0"
dash-bootstrap-components = ">=1.6.0,<1.7.0"
fastapi = ">=0.115.5,<0.116.0"
uvicorn = ">=0.32.1,<0.33.0"
flask-caching = ">=2.3.0,<2.4.0"
folium = ">=0.18.0,<0.19.0"
selenium = ">=4.27.1,<4.28.0"
imgkit = ">=1.2.3,<1.3.0"
wkhtmltopdf = "^0.2"
jupyter-cache = ">=1.0.1,<1.1.0"
openai = "^1.59.6"
fastapi-cache2 = ">=0.2.2,<0.3.0"
geopandas = "^1.0.1"
poetry = "^2.0.1"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
psycopg2-binary = ">=2.9.10,<2.10.0"
psycopg2 = ">=2.9.10,<2.10.0"
ujson = ">=5.10.0,<5.11.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
